[docs] Use a local .json file for contributors. (#6076)
* Use a local .json file for contributors.

* Take doc slugs into account.
