+++ developer-docs-site/docusaurus.config.js
@@ -5,7 +5,6 @@ const lightCodeTheme = require("prism-react-renderer/themes/github");
 const darkCodeTheme = require("prism-react-renderer/themes/dracula");
 
 const codeInjector = require("./src/remark/code-injector");
-const contributors = require("./src/remark/contributors");
 
 const { ProvidePlugin } = require("webpack");
 
@@ -35,7 +34,7 @@ const config = {
           sidebarPath: require.resolve("./sidebars.js"),
           sidebarCollapsible: false,
           editUrl: "https://github.com/aptos-labs/aptos-core/tree/main/developer-docs-site/",
-          remarkPlugins: [codeInjector, math, contributors],
+          remarkPlugins: [codeInjector, math],
           path: "docs",
           rehypePlugins: [katex],
         },
