+++ developer-docs-site/package.json
@@ -15,7 +15,8 @@
     "write-heading-ids": "docusaurus write-heading-ids",
     "typecheck": "tsc",
     "fmt": "prettier --write '**/*.(tsx|ts|js)'",
-    "lint": "pnpm install --frozen-lockfile --silent && prettier --check '**/*.(tsx|ts|js)'"
+    "lint": "pnpm install --frozen-lockfile --silent && prettier --check '**/*.(tsx|ts|js)'",
+    "contributors": "scripts/contributors.js"
   },
   "dependencies": {
     "@algolia/client-search": "^4.14.2",
